% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_draw_inputs}
\alias{check_draw_inputs}
\title{Check the inputs to the draw function}
\usage{
check_draw_inputs(draws, nDraws, draw_dimensions, Nindividuals)
}
\arguments{
\item{draws}{The specified Model}

\item{nDraws}{Named vector of proposed start values for the model}

\item{draw_dimensions}{the dataset on which to estimate}

\item{Nindividuals}{The availabilities for the alternatives in the model specification}
}
\value{
A list consisting of the checked draws and Ndraws, both computed if required)
}
\description{
Check the inputs to the draw function
}
