% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{check_inputs}
\alias{check_inputs}
\title{Check the inputs to the estimate function}
\usage{
check_inputs(model_spec, start_values, data, availabilities, draws, fixedparam)
}
\arguments{
\item{model_spec}{The specified Model}

\item{start_values}{Named vector of proposed start values for the model}

\item{data}{the dataset on which to estimate}

\item{availabilities}{The availabilities for the alternatives in the model specification}

\item{draws}{The matrix of random draws}

\item{fixedparam}{Named vector of parameters to be fixed}
}
\value{
Nothing
}
\description{
This function checks the start_vlaues, data, availabilities, draws and fixedparams for validity.
If this function runs without error, then the inputs are valid for the maxLikelihood function.
These checks are important, because an error in the internal C++ code will cause the Rstudio session to crash.
Incidentally, if there is concern of this happening, it is recommended to run the script from the
command line, using Rscript.
}
