% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{compileUtilityFunction}
\alias{compileUtilityFunction}
\title{compileUtilityFunction
Deprecated, please see \code{\link[=specify_model]{specify_model()}}}
\usage{
compileUtilityFunction(...)
}
\arguments{
\item{...}{Parameters to specify_model}
}
\description{
compileUtilityFunction
Deprecated, please see \code{\link[=specify_model]{specify_model()}}
}
