% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_halton_draws}
\alias{create_halton_draws}
\title{Create a standard set of Halton draws to use in estimation}
\usage{
create_halton_draws(Nindividuals, nDraws, draw_dimensions)
}
\arguments{
\item{Nindividuals}{The number individuals in the dataset}

\item{nDraws}{The number of draws needed}

\item{draw_dimensions}{the number of draw dimensions needed}
}
\value{
Matrix of availabilities for alternatives and the number of choice observations
}
\description{
Create a standard set of Halton draws to use in estimation
}
\examples{
create_halton_draws(100, 10, 5)
create_halton_draws(100, 100, 20)

}
