% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_default_availabilities}
\alias{generate_default_availabilities}
\title{Generate a ones-matrix of availabilities}
\usage{
generate_default_availabilities(data, num_utility_functions)
}
\arguments{
\item{data}{The dataset used in the model}

\item{num_utility_functions}{the number of alternatives in the model}
}
\value{
Ones-matrix of availabilities for alternatives and the number of choice observations
}
\description{
Generate a ones-matrix of availabilities
}
\examples{
data("Train", package="mlogit")
Train$ID <- Train$id
Train$CHOICE <- as.numeric(Train$choice)
generate_default_availabilities(Train, 5)
}
