% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_posterior.R
\name{posteriors}
\alias{posteriors}
\title{Calculate the posteriors for a specified and estimated model}
\usage{
posteriors(model, indiv_data = NULL, code_output_file = NULL)
}
\arguments{
\item{model}{The estimated Model}

\item{indiv_data}{Alternative individual data to use insteaf of that in the dataset}

\item{code_output_file}{An (optional) location where the compiled code should be saved (useful for debugging}
}
\value{
Dataframe of individual-level posteriors
}
\description{
Calculate the posteriors for a specified and estimated model
}
