% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{probabilities}
\alias{probabilities}
\title{Calculate the probabilities for a specified and estimated model. 
Note that if new data or draws are provided, the model will not be re-estimated}
\usage{
probabilities(model, data = NULL, availabilities = NULL,
  draws = NULL, nDraws = NULL, num_threads = 1)
}
\arguments{
\item{model}{The estimated Model}

\item{data}{(Optional) New data to use insteaf of that in the dataset}

\item{availabilities}{(Optional) New availabilites to use}

\item{draws}{(Optional) Optional new set of random draws to use}

\item{nDraws}{(Optional) Optional new number of random draws to use}

\item{num_threads}{Enable parallel computing where available using this many cores}
}
\value{
Dataframe of individual-level posteriors
}
\description{
Calculate the probabilities for a specified and estimated model. 
Note that if new data or draws are provided, the model will not be re-estimated
}
