% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{summary_tex}
\alias{summary_tex}
\title{Return tex formatted output of a model summary}
\usage{
summary_tex(model_summary)
}
\arguments{
\item{model_summary}{A summary of an estimated Model}
}
\value{
Formatted text output suitable for a research paper.
}
\description{
Return tex formatted output of a model summary
}
